Update example project structure

+++ Example/Paged Tabs Example/Screens/ChildViewController.m
@@ -17,7 +17,7 @@
 - (void)viewDidLoad {
     [super viewDidLoad];
     
-    self.titleLabel.text = [NSString stringWithFormat:@"Page %i", self.pageIndex + 1];
+    self.titleLabel.text = [NSString stringWithFormat:@"Page %i", (int)(self.pageIndex + 1)];
 }
 
 @end
